- renaming request.py to requests.py - changed HSV function to pascal case - fixed function calling in requests.py
